Minor changes in alacritty install script
- Fixed type that would try to remove the wrong package when removing
snapd from the system
- added statement to check for the nosnap.pref file before removing it,
as it would result in an error otherwise (if the file wouldn't exist)
- removed downloading the statement that checks for the dotfiles repo,
as there is not reason why it would not exist

+++ scripts/install-alacritty.sh
@@ -10,7 +10,10 @@ mkdir -p ~/.config/alacritty/
 sudo apt install -y cmake pkg-config libfreetype6-dev libfontconfig1-dev libxcb-xfixes0-dev python3 git
 
 # install the snapstore to get rustup
-sudo rm /etc/apt/preferences.d/nosnap.pref
+if [ -f /etc/apt/preferences.d/nosnap.pref ] ; then
+    sudo rm /etc/apt/preferences.d/nosnap.pref
+fi
+
 sudo apt install -y snapd && sudo snap install rustup --classic
 
 rustup override set stable && rustup update stable
@@ -24,18 +27,13 @@ cd alacritty
 cargo build --release && sudo cp target/release/alacritty /usr/local/bin/
 
 # removing all snap stuff cause it sucks
-sudo apt purge snap -y
+sudo apt purge snapd -y
 rm -rf ~/snap
 
 cd ~
 
 # cleaning home directory
 rm -rf alacritty
-
-if ! [ -r "~/dotfiles" ]
-then
-    git clone https://github.com/TomSchammo/dotfiles.git
-fi
 
 # copying config into the right directory
 mv ~/dotfiles/dotfiles/alacritty.yml ~/.config/alacritty/alacritty.yml
